+++ airline-core/src/test/java/com/github/rvesse/airline/TestGitByAnnotation.java
@@ -16,13 +16,17 @@
 package com.github.rvesse.airline;
 
 import org.apache.commons.lang3.StringUtils;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class TestGitByAnnotation
-{
+import com.github.rvesse.airline.builder.ParserBuilder;
+import com.github.rvesse.airline.help.Help;
+import com.github.rvesse.airline.model.ParserMetadata;
+import com.github.rvesse.airline.parser.errors.ParseCommandUnrecognizedException;
+
+public class TestGitByAnnotation {
     @Test
-    public void test()
-    {
+    public void test() {
         // simple command parsing example
         git("add", "-p", "file");
         git("remote", "add", "origin", "git@github.com:airlift/airline.git");
@@ -41,11 +45,58 @@ public class TestGitByAnnotation
         git("help", "remote");
         git("help", "remote", "show");
     }
-    
-    private void git(String... args)
-    {
+
+    @Test
+    public void testParserConfigOverride() {
+        ParserMetadata<Runnable> parserConfig = new ParserBuilder<Runnable>().withCommandAbbreviation().build();
+
+        git(parserConfig, "rem");
+        git(parserConfig, "rem show");
+        git(parserConfig, "rem sh");
+        git(parserConfig, "remote sh");
+    }
+
+    @Test
+    public void testParserConfigOverride2() {
+        ParserMetadata<Runnable> parserConfig = new ParserBuilder<Runnable>().withCommandAbbreviation().build();
+
+        // This works because the base config includes the alias foo
+        gitOverridden(null, "foo");
+
+        // These all work because our override enabled command abbreviation
+        gitOverridden(parserConfig, "rem");
+        gitOverridden(parserConfig, "rem", "show");
+        gitOverridden(parserConfig, "rem", "sh");
+        gitOverridden(parserConfig, "remote", "sh");
+
+        // This returns help because the alias definition from the base config is overridden by the provided parser config
+        com.github.rvesse.airline.Cli<Runnable> gitParser = new com.github.rvesse.airline.Cli<Runnable>(
+                GitWithCliAnnotation2.class, parserConfig);
+        Runnable cmd = gitParser.parse("foo");
+        Assert.assertEquals(cmd.getClass(), Help.class);
+    }
+
+    private void git(String... args) {
         System.out.println("$ git " + StringUtils.join(args, ' '));
         GitWithCliAnnotation.run(args);
+        System.out.println();
+    }
+
+    private void git(ParserMetadata<Runnable> parserConfig, String... args) {
+        System.out.println("$ git " + StringUtils.join(args, ' '));
+        com.github.rvesse.airline.Cli<Runnable> gitParser = new com.github.rvesse.airline.Cli<Runnable>(
+                GitWithCliAnnotation.class, parserConfig);
+
+        gitParser.parse(args).run();
+        System.out.println();
+    }
+
+    private void gitOverridden(ParserMetadata<Runnable> parserConfig, String... args) {
+        System.out.println("$ git " + StringUtils.join(args, ' '));
+        com.github.rvesse.airline.Cli<Runnable> gitParser = new com.github.rvesse.airline.Cli<Runnable>(
+                GitWithCliAnnotation2.class, parserConfig);
+
+        gitParser.parse(args).run();
         System.out.println();
     }
 }